Move unit-tests and fix pep8
